// admin-panel/src/types/Notifications.ts
export interface NotificationData {
    ads: string;
    body: string;
    broker_package: string;
    broker_request: string;
    createdAt: Date;
    featuredProperty: string;
    payment: string;
    priority: string;
    property: string;
    readAt: Date | null;
    recipients: string[];
    report: string;
    sender: string;
    title: string;
    _id: string;
    user: string;
  }
  